Low-light Image Enhancement
A Convolutional Neural Network (CNN) directly learns an end-to-end mapping between low light and bright images. The low-light image enhancement in this code was solved as a machine-learning problem using the TensorFlow library. This model considered different metrics  SSIM, PSNR and entropy as the loss functions to improve the contrast of low and hazy images.